+++ packages/you_flutter/lib/src/conventions.dart
@@ -1,0 +1,9 @@
+///  Configuration conventions for the note package.
+library;
+
+Conventions conventions = Conventions._();
+
+/// 惯用规则
+class Conventions {
+  Conventions._();
+}
