+++ packages/you_flutter/lib/src/note/contents/outline.dart
@@ -1,5 +1,6 @@
 
 // markdown 的结构轮廓，主要用来显示TOC
+import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:meta/meta.dart';
 
@@ -19,6 +20,21 @@ class Outline {
     current = current!.add(newNode);
   }
 
+  // @override
+  // void initState() {
+  //   super.initState();
+  //
+  //   WidgetsBinding.instance.addPostFrameCallback((_) {
+  //     // to do
+  //     // flutter-markdown只有在Widget.build时才parse markdown，导致第一次[build]时,
+  //     // 装配的outline无法展示出来， 所以需要触发第二次build,以使其展示出来
+  //     // 暂时没想好最终处理办法，暂时这样。
+  //     setState(() {
+  //       outline.collectDone();
+  //     });
+  //   });
+  // }
+  //
   /// bed design: 目前非常糟糕的设计，因为outline会在markdown 第一次Widget.build后才能装配好
   /// 第一次build时 界面上是看不到outline的，后面如果因为resize多次build，会造成outline持续重复增加内容
   /// 所以要结束掉它
@@ -76,5 +92,62 @@ class OutlineNode {
 
   void clear() {
     children.clear();
+  }
+}
+
+
+
+class OutlineTreeView extends StatelessWidget {
+  final Outline outline;
+
+  // 主内容部分的滚动控制，点击outline触发主屏滚动到指定标题
+  final ScrollController pageController;
+
+  const OutlineTreeView({required this.outline, required this.pageController});
+
+  @override
+  Widget build(BuildContext context) {
+    // 一页一个链接
+    Widget headLink(OutlineNode node) {
+      var link2 = TextButton(
+        style: ButtonStyle(padding: WidgetStateProperty.all<EdgeInsets>(const EdgeInsets.all(2))),
+        child: Row(
+          crossAxisAlignment: CrossAxisAlignment.center,
+          children: [
+            // const Icon(
+            //   Icons.arrow_right,
+            // ),
+            // title 被Flexible包裹后，文本太长会自动换行▽
+            // 换行后左边图标需要CrossAxisAlignment.start 排在文本的第一行
+            //📜📁📂📄🗓📜 ▸▾▹▿▶︎▷▼▽►🔘◽️▫️◻️◼️⬛️🔹⚉
+            Flexible(child: Text("◻ ${node.title}")),
+          ],
+        ),
+        onPressed: () {
+          // 防止异常
+          if (pageController.hasClients) {
+            Scrollable.ensureVisible(node.key.currentContext!);
+          }
+        },
+      );
+      // TextButton link = TextButton(onPressed: (){}, child: Text(node.title));
+      return Padding(
+        // 缩进模仿树形
+        padding: EdgeInsets.only(left: 20 * (node.level - 1).toDouble()),
+        child: link2,
+      );
+    }
+
+    var nodes = outline.root.toList(includeThis: false);
+
+    var column = Column(
+      children: [
+        ...nodes.map((e) => headLink(e)),
+      ],
+    );
+    return SingleChildScrollView(
+      scrollDirection: Axis.vertical,
+      child: column,
+    );
   }
 }
