+++ packages/you_flutter/lib/src/note/contents/source_code.dart
@@ -9,7 +9,7 @@ import 'package:analyzer/file_system/memory_file_system.dart';
 import 'package:analyzer/src/test_utilities/mock_sdk.dart';
 import 'package:meta/meta.dart';
 import 'package:path/path.dart' as path_;
-import 'package:you_flutter/src/note/utils_core.dart';
+import 'package:you_flutter/src/utils_core.dart';
 // ignore: implementation_imports, internal import /src of the same project
 import 'package:you_dart/src/core.dart';
