+++ packages/you_flutter/lib/src/note/note.dart
@@ -8,8 +8,6 @@ import 'package:you_flutter/router.dart';
 import 'package:you_flutter/src/note/contents/contents.dart';
 import 'package:you_flutter/src/router.dart';
 import 'package:you_flutter/state.dart';
-import 'package:you_flutter/src/note/note_conf.dart';
-import 'package:you_flutter/src/note/conventions.dart';
 import 'package:http/http.dart' as http;
 
 typedef NoteBuilder = void Function(BuildContext context, Cell print);
@@ -60,135 +58,6 @@ final class _DefaultNote extends StatelessWidget with NoteMixin {
           }).toList(),
         ));
   }
-}
-
-@Deprecated("已被you_router取代，待删除")
-class NoteRoute {
-  /// A file system term,  that refers to the last part of a path
-  /// example: a/b/c , c is basename
-  final String basename;
-  final Map<String, NoteRoute> _children = {};
-  final NoteRoute? parent;
-  bool expand = false;
-
-  NoteConf? conf;
-
-  NoteRoute.root()
-      : basename = "",
-        parent = null;
-
-  void configTree({int extendLevel = 0}) {
-    if (extendLevel <= 0) return;
-
-    expand = extendLevel > 0;
-    for (var e in children) {
-      e.configTree(extendLevel: extendLevel - 1);
-    }
-  }
-
-  List<NoteRoute> get children => List.from(_children.values);
-
-  bool get isLeaf => _children.isEmpty;
-
-  int get level => isRoot ? 0 : parent!.level + 1;
-
-  int levelTo(NoteRoute parent) => level - parent.level;
-
-  List<NoteRoute> get ancestors => isRoot ? [] : [parent!, ...parent!.ancestors];
-
-  List<NoteRoute> get meAndAncestors => [this, ...ancestors];
-
-  bool get isRoot => parent == null;
-
-  NoteRoute get root => isRoot ? this : parent!.root;
-
-  /// Note names, which can be set to human-readable names in note.json,
-  /// are displayed on the navigation tree
-  String get displayName => conf != null ? conf!.displayName : basename;
-
-  String get path {
-    if (isRoot) return "/";
-    var parentPath = parent!.path;
-    return parentPath == "/" ? "/$basename" : "$parentPath/$basename";
-  }
-
-  List<NoteRoute> toList({
-    bool includeThis = true,
-    bool Function(NoteRoute path)? test,
-    Comparator<NoteRoute>? sortBy,
-  }) {
-    test = test ?? (e) => true;
-    if (!test(this)) {
-      return [];
-    }
-    List<NoteRoute> children = List.from(_children.values);
-    if (sortBy != null) {
-      children.sort(sortBy);
-    }
-
-    var flatChildren = children.expand((child) {
-      return child.toList(includeThis: true, test: test, sortBy: sortBy);
-    }).toList();
-    return includeThis ? [this, ...flatChildren] : flatChildren;
-  }
-
-  List<NoteRoute> topList({bool topDown = true}) {
-    if (isRoot) {
-      return [this];
-    }
-    return [...parent!.topList(), this];
-  }
-
-  // todo bug
-  NoteRoute? child(String path) {
-    NoteRoute? result = this;
-    for (var split in path.split("/").map((e) => e.trim()).where((e) => e != "")) {
-      result = result?._children[split];
-      if (result == null) break;
-    }
-    return result;
-  }
-
-  /// 扁平化name，去掉排序用的数字前缀
-  String get nameFlat {
-    return basename.replaceAll(RegExp("\\d+[.]"), "") // 1.z.about -> z.about
-        ;
-  }
-
-  String toStringShort() {
-    return path;
-  }
-
-  @override
-  String toString({bool? deep}) {
-    if (deep == null || !deep) {
-      return "Page($path ,kids:${_children.values.map((e) => e.toStringShort()).toList()})";
-    } else {
-      StringBuffer sb = StringBuffer();
-      for (NoteRoute n in toList()) {
-        sb.write("$n\n");
-      }
-      return sb.toString();
-    }
-  }
-
-  bool contains(String location) {
-    var c = child(location);
-    return c != null;
-  }
-
-  void visit(bool Function(NoteRoute note) visitor) {
-    if (!visitor(this)) {
-      return;
-    }
-    for (var child in _children.values) {
-      child.visit(visitor);
-    }
-  }
-
-  String get dartAssetPath => conventions.noteDartAssetPath(path);
-
-  String get confAssetPath => conventions.noteConfAssetPath(path);
 }
 
 base class Cell {
