+++ packages/you_flutter/pubspec.yaml
@@ -11,6 +11,7 @@ environment:
 dependencies:
   code_builder: ^4.10.0
   dart_style: ^2.3.6
+  file: ^7.0.0
   flutter:
     sdk: flutter
   flutter_highlight: ^0.7.0
